+++ src/model.rs
@@ -3,7 +3,23 @@ use std::path::Path;
 use std::process::Command;
 
 use chrono::Utc;
+use clap::ValueEnum;
 use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Eq, ValueEnum)]
+pub enum Priority {
+    // clap default to lower case, so add aliases for upper case too
+    #[value(alias = "P0")]
+    P0,
+    #[value(alias = "P1")]
+    P1,
+    #[value(alias = "P2")]
+    P2,
+    #[value(alias = "P3")]
+    P3,
+    #[value(alias = "P4")]
+    P4,
+}
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Meta {
@@ -14,6 +30,7 @@ pub struct Meta {
     pub type_: String,
     pub labels: Vec<String>,
     pub assignee: String,
+    pub priority: Priority,
     pub created: String,
     pub updated: String,
 }